Rename package: rename Duke package to zhen.
Java coding standard says package names should be in lowercase.

Let's rename the package name from Duke to zhen.

Changing from Duke to zhen,
* follows coding standard
* shows name of my program
