+++ src/main/java/zhen/DialogBox.java
@@ -1,4 +1,4 @@
-package Duke;
+package zhen;
 
 import java.io.IOException;
 import java.util.Collections;
