+++ src/main/java/zhen/Launcher.java
@@ -1,8 +1,5 @@
-package Duke;
+package zhen;
 import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Label;
-import javafx.stage.Stage;
 
 public class Launcher {
     public static void main(String[] args) {
