+++ src/main/java/zhen/Main.java
@@ -1,4 +1,4 @@
-package Duke;
+package zhen;
 
 import java.io.IOException;
 
@@ -13,7 +13,7 @@ import javafx.stage.Stage;
  */
 public class Main extends Application {
 
-    private Duke duke = new Duke("database.ser");
+    private Zhen duke = new Zhen("database.ser");
 
     @Override
     public void start(Stage stage) {
@@ -23,7 +23,7 @@ public class Main extends Application {
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
+            fxmlLoader.<MainWindow>getController().setZhen(duke);
             stage.show();
         } catch (IOException e) {
             e.printStackTrace();
