+++ src/main/java/zhen/Storage.java
@@ -1,8 +1,8 @@
-package Duke;
+package zhen;
 import java.io.*;
 import java.util.ArrayList;
 
-import Duke.task.*;
+import zhen.task.*;
 
 /**
  * Class responsible for storing task information locally and restoring the chat from local
