+++ src/main/java/zhen/command/DontknowCommand.java
@@ -1,19 +1,18 @@
-package Duke.command;
-import Duke.*;
+package zhen.command;
+import zhen.*;
 
 /**
- * Represent command of listing all the task added by the user.
+ * Represent command of showing user that this program don't understand s/he command.
  */
-public class ListCommand extends Command {
+public class DontknowCommand extends Command {
     /**
-     * List all the task stored in the program.
-     * @param taskList the task list storing the task the user operated on.
+     * Display message to user to tell that the program don't know what user's input means
+     * @param taskList the task list potentially be worked on, not used in this program
      * @param ui to manage the interaction with users
      * @param storage to store the product of execution locally, not used in this program
      */
-    @Override
     public String execute(TaskList taskList, Ui ui, Storage storage) {
-        String replyMessage = taskList.toString();
+        String replyMessage = "OOPS!!! I'm sorry, but I don't know what that means";
         Ui.print_message(replyMessage);
         return replyMessage;
     }
