+++ src/main/java/zhen/command/ExitCommand.java
@@ -1,5 +1,5 @@
-package Duke.command;
-import Duke.*;
+package zhen.command;
+import zhen.*;
 /**
  * Represent the command of exiting the application
  */
