+++ src/main/java/zhen/task/Deadline.java
@@ -1,4 +1,4 @@
-package Duke.task;
+package zhen.task;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.Locale;
