+++ src/main/java/zhen/task/Event.java
@@ -1,4 +1,4 @@
-package Duke.task;
+package zhen.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
