+++ src/main/java/zhen/task/Task.java
@@ -1,4 +1,4 @@
-package Duke.task;
+package zhen.task;
 
 import java.io.Serializable;
