+++ src/test/java/zhen/ParserTest.java
@@ -1,4 +1,4 @@
-package Duke;
+package zhen;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
